Module hotshot_task_impls::consensus::handlers

Helper functions to handle proposal-related functionality.

Functions

create_and_send_proposal

Create the header for a proposal, build the proposal, and broadcast the proposal send evnet.

handle_quorum_proposal_recv 🔒

Handle the received quorum proposal.

handle_quorum_proposal_validated

Handle QuorumProposalValidated event content and submit a proposal if possible.

publish_proposal_from_commitment_and_metadata

Send a proposal for the view view from the latest high_qc given an upgrade cert. This is the standard case proposal scenario.

publish_proposal_if_able

Publishes a proposal if there exists a value which we can propose from. Specifically, we must have either commitment_and_metadata, or a decided_upgrade_cert.

update_state_and_vote_if_able

Check if we are able to vote, like whether the proposal is valid, whether we have DAC and VID share, and if so, vote.

Type Aliases

VoteInfo 🔒

Private key, latest decided upgrade certificate, committee membership, and event stream, for sending the vote.