hotshot_task_impls::consensus::handlers

Function send_high_qc

pub async fn send_high_qc<TYPES: NodeType, V: Versions, I: NodeImplementation<TYPES>>(
    new_view_number: TYPES::View,
    sender: &Sender<Arc<HotShotEvent<TYPES>>>,
    receiver: &Receiver<Arc<HotShotEvent<TYPES>>>,
    task_state: &mut ConsensusTaskState<TYPES, I, V>,
) -> Result<()>

Send an event to the next leader containing the highest QC we have This is a necessary part of HotStuff 2 but not the original HotStuff

#Errors Returns and error if we can’t get the version or the version doesn’t yet support HS 2