Module hotshot_task_impls::consensus::helpers

Helper functions to handle proposal-related functionality.

Structs

LeafChainTraversalOutcome

Helper type to give names and to the output values of the leaf chain traversal operation.

Functions

create_and_send_proposal

Create the header for a proposal, build the proposal, and broadcast the proposal send evnet.

decide_from_proposal

Ascends the leaf chain by traversing through the parent commitments of the proposal. We begin by obtaining the parent view, and if we are in a chain (i.e. the next view from the parent is one view newer), then we begin attempting to form the chain. This is a direct impl from HotStuff section 5:

fetch_proposal 🔒

Trigger a request to the network for a proposal for a view and wait for the response

handle_quorum_proposal_recv 🔒

Handle the received quorum proposal.

handle_quorum_proposal_validated

Handle QuorumProposalValidated event content and submit a proposal if possible.

parent_leaf_and_state 🔒

Gets the parent leaf and state from the parent of a proposal, returning an anyhow::Error if not.

publish_proposal_from_commitment_and_metadata

Send a proposal for the view view from the latest high_qc given an upgrade cert. This is the standard case proposal scenario.

publish_proposal_if_able

Publishes a proposal if there exists a value which we can propose from. Specifically, we must have either commitment_and_metadata, or a decided_upgrade_cert.

temp_validate_proposal_safety_and_liveness

Validate the state and safety and liveness of a proposal then emit a QuorumProposalValidated event.

update_state_and_vote_if_able

Check if we are able to vote, like whether the proposal is valid, whether we have DAC and VID share, and if so, vote.

validate_proposal_safety_and_liveness

Validate the state and safety and liveness of a proposal then emit a QuorumProposalValidated event.

validate_proposal_view_and_certs

Validates, from a given proposal that the view that it is being submitted for is valid when compared to cur_view which is the highest proposed view (so far) for the caller. If the proposal is for a view that’s later than expected, that the proposal includes a timeout or view sync certificate.

Type Aliases

VoteInfo 🔒

Private key, latest decided upgrade certificate, committee membership, and event stream, for sending the vote.