Module hotshot_task_impls::consensus::proposal_helpers

Helper functions to handler proposal-related functionality.

Functions

create_and_send_proposal

Create the header for a proposal, build the proposal, and broadcast the proposal send evnet.

get_parent_leaf_and_state

Gets the parent leaf and state from the parent of a proposal, returning an anyhow::Error if not.

handle_quorum_proposal_recv

Handle the received quorum proposal.

publish_proposal_from_commitment_and_metadata 🔒

Send a proposal for the view view from the latest high_qc given an upgrade cert. This is the standard case proposal scenario.

publish_proposal_from_upgrade_cert 🔒

Send a proposal for the view view from the latest high_qc given an upgrade cert. This is a special case proposal scenario.

publish_proposal_if_able

Publishes a proposal if there exists a value which we can propose from. Specifically, we must have either commitment_and_metadata, or a decided_upgrade_cert.

validate_proposal_safety_and_liveness

Validate the state and safety and liveness of a proposal then emit a QuorumProposalValidated event.

validate_proposal_view_and_certs

Validates, from a given proposal that the view that it is being submitted for is valid when compared to cur_view which is the highest proposed view (so far) for the caller. If the proposal is for a view that’s later than expected, that the proposal includes a timeout or view sync certificate.