Module hotshot_task_impls::consensus::view_change

Handles view-change related functionality.

Constants

DONT_SEND_VIEW_CHANGE_EVENT 🔒

Constant which tells update_view to not send a view change event when called.

SEND_VIEW_CHANGE_EVENT 🔒

Constant which tells update_view to send a view change event when called.

Functions

update_view 🔒

Update the view if it actually changed, takes a mutable reference to the cur_view and the timeout_task which are updated during the operation of the function.