Enum hotshot_task_impls::events::HotShotEvent

pub enum HotShotEvent<TYPES: NodeType> {

    Shutdown,
    QuorumProposalRecv(Proposal<TYPES, QuorumProposalWrapper<TYPES>>, TYPES::SignatureKey),
    QuorumVoteRecv(QuorumVote2<TYPES>),
    TimeoutVoteRecv(TimeoutVote2<TYPES>),
    TimeoutVoteSend(TimeoutVote2<TYPES>),
    DaProposalRecv(Proposal<TYPES, DaProposal2<TYPES>>, TYPES::SignatureKey),
    DaProposalValidated(Proposal<TYPES, DaProposal2<TYPES>>, TYPES::SignatureKey),
    DaVoteRecv(DaVote2<TYPES>),
    DaCertificateRecv(DaCertificate2<TYPES>),
    DaCertificateValidated(DaCertificate2<TYPES>),
    QuorumProposalSend(Proposal<TYPES, QuorumProposalWrapper<TYPES>>, TYPES::SignatureKey),
    QuorumVoteSend(QuorumVote2<TYPES>),
    ExtendedQuorumVoteSend(QuorumVote2<TYPES>),
    QuorumProposalValidated(Proposal<TYPES, QuorumProposalWrapper<TYPES>>, Leaf2<TYPES>),
    QuorumProposalRequestSend(ProposalRequestPayload<TYPES>, <TYPES::SignatureKey as SignatureKey>::PureAssembledSignatureType),
    QuorumProposalRequestRecv(ProposalRequestPayload<TYPES>, <TYPES::SignatureKey as SignatureKey>::PureAssembledSignatureType),
    QuorumProposalResponseSend(TYPES::SignatureKey, Proposal<TYPES, QuorumProposalWrapper<TYPES>>),
    QuorumProposalResponseRecv(Proposal<TYPES, QuorumProposalWrapper<TYPES>>),
    DaProposalSend(Proposal<TYPES, DaProposal2<TYPES>>, TYPES::SignatureKey),
    DaVoteSend(DaVote2<TYPES>),
    QcFormed(Either<QuorumCertificate<TYPES>, TimeoutCertificate<TYPES>>),
    Qc2Formed(Either<QuorumCertificate2<TYPES>, TimeoutCertificate2<TYPES>>),
    NextEpochQc2Formed(Either<NextEpochQuorumCertificate2<TYPES>, TimeoutCertificate<TYPES>>),
    DacSend(DaCertificate2<TYPES>, TYPES::SignatureKey),
    ViewChange(TYPES::View, Option<TYPES::Epoch>),
    ViewSyncTimeout(TYPES::View, u64, ViewSyncPhase),
    ViewSyncPreCommitVoteRecv(ViewSyncPreCommitVote2<TYPES>),
    ViewSyncCommitVoteRecv(ViewSyncCommitVote2<TYPES>),
    ViewSyncFinalizeVoteRecv(ViewSyncFinalizeVote2<TYPES>),
    ViewSyncPreCommitVoteSend(ViewSyncPreCommitVote2<TYPES>),
    ViewSyncCommitVoteSend(ViewSyncCommitVote2<TYPES>),
    ViewSyncFinalizeVoteSend(ViewSyncFinalizeVote2<TYPES>),
    ViewSyncPreCommitCertificateRecv(ViewSyncPreCommitCertificate2<TYPES>),
    ViewSyncCommitCertificateRecv(ViewSyncCommitCertificate2<TYPES>),
    ViewSyncFinalizeCertificateRecv(ViewSyncFinalizeCertificate2<TYPES>),
    ViewSyncPreCommitCertificateSend(ViewSyncPreCommitCertificate2<TYPES>, TYPES::SignatureKey),
    ViewSyncCommitCertificateSend(ViewSyncCommitCertificate2<TYPES>, TYPES::SignatureKey),
    ViewSyncFinalizeCertificateSend(ViewSyncFinalizeCertificate2<TYPES>, TYPES::SignatureKey),
    ViewSyncTrigger(TYPES::View),
    Timeout(TYPES::View, Option<TYPES::Epoch>),
    TransactionsRecv(Vec<TYPES::Transaction>),
    TransactionSend(TYPES::Transaction, TYPES::SignatureKey),
    SendPayloadCommitmentAndMetadata(VidCommitment, BuilderCommitment, <TYPES::BlockPayload as BlockPayload<TYPES>>::Metadata, TYPES::View, Vec1<BuilderFee<TYPES>>, Option<TYPES::AuctionResult>),
    BlockRecv(PackedBundle<TYPES>),
    VidDisperseSend(Proposal<TYPES, VidDisperse<TYPES>>, TYPES::SignatureKey),
    VidShareRecv(TYPES::SignatureKey, Proposal<TYPES, VidDisperseShare2<TYPES>>),
    VidShareValidated(Proposal<TYPES, VidDisperseShare2<TYPES>>),
    UpgradeProposalRecv(Proposal<TYPES, UpgradeProposal<TYPES>>, TYPES::SignatureKey),
    UpgradeProposalSend(Proposal<TYPES, UpgradeProposal<TYPES>>, TYPES::SignatureKey),
    UpgradeVoteRecv(UpgradeVote<TYPES>),
    UpgradeVoteSend(UpgradeVote<TYPES>),
    UpgradeCertificateFormed(UpgradeCertificate<TYPES>),
    QuorumProposalPreliminarilyValidated(Proposal<TYPES, QuorumProposalWrapper<TYPES>>),
    VidRequestSend(DataRequest<TYPES>, TYPES::SignatureKey, TYPES::SignatureKey),
    VidRequestRecv(DataRequest<TYPES>, TYPES::SignatureKey),
    VidResponseSend(TYPES::SignatureKey, TYPES::SignatureKey, Proposal<TYPES, VidDisperseShare2<TYPES>>),
    VidResponseRecv(TYPES::SignatureKey, Proposal<TYPES, VidDisperseShare2<TYPES>>),
    HighQcRecv(QuorumCertificate2<TYPES>, TYPES::SignatureKey),
    HighQcSend(QuorumCertificate2<TYPES>, TYPES::SignatureKey, TYPES::SignatureKey),

}

All of the possible events that can be passed between Sequencing HotShot tasks

Variants

Shutdown

Shutdown the task

QuorumProposalRecv(Proposal<TYPES, QuorumProposalWrapper<TYPES>>, TYPES::SignatureKey)

A quorum proposal has been received from the network; handled by the consensus task

QuorumVoteRecv(QuorumVote2<TYPES>)

A quorum vote has been received from the network; handled by the consensus task

TimeoutVoteRecv(TimeoutVote2<TYPES>)

A timeout vote received from the network; handled by consensus task

TimeoutVoteSend(TimeoutVote2<TYPES>)

Send a timeout vote to the network; emitted by consensus task replicas

DaProposalRecv(Proposal<TYPES, DaProposal2<TYPES>>, TYPES::SignatureKey)

A DA proposal has been received from the network; handled by the DA task

DaProposalValidated(Proposal<TYPES, DaProposal2<TYPES>>, TYPES::SignatureKey)

A DA proposal has been validated; handled by the DA task and VID task

DaVoteRecv(DaVote2<TYPES>)

A DA vote has been received by the network; handled by the DA task

DaCertificateRecv(DaCertificate2<TYPES>)

A Data Availability Certificate (DAC) has been received by the network; handled by the consensus task

DaCertificateValidated(DaCertificate2<TYPES>)

A DAC is validated.

QuorumProposalSend(Proposal<TYPES, QuorumProposalWrapper<TYPES>>, TYPES::SignatureKey)

Send a quorum proposal to the network; emitted by the leader in the consensus task

QuorumVoteSend(QuorumVote2<TYPES>)

Send a quorum vote to the next leader; emitted by a replica in the consensus task after seeing a valid quorum proposal

ExtendedQuorumVoteSend(QuorumVote2<TYPES>)

Broadcast a quorum vote to form an eQC; emitted by a replica in the consensus task after seeing a valid quorum proposal

QuorumProposalValidated(Proposal<TYPES, QuorumProposalWrapper<TYPES>>, Leaf2<TYPES>)

A quorum proposal with the given parent leaf is validated. The full validation checks include:

The proposal is not for an old view

The proposal has been correctly signed by the leader of the current view

The justify QC is valid

The proposal passes either liveness or safety check.

QuorumProposalRequestSend(ProposalRequestPayload<TYPES>, <TYPES::SignatureKey as SignatureKey>::PureAssembledSignatureType)

A quorum proposal is missing for a view that we need.

QuorumProposalRequestRecv(ProposalRequestPayload<TYPES>, <TYPES::SignatureKey as SignatureKey>::PureAssembledSignatureType)

A quorum proposal was requested by a node for a view.

QuorumProposalResponseSend(TYPES::SignatureKey, Proposal<TYPES, QuorumProposalWrapper<TYPES>>)

A quorum proposal was missing for a view. As the leader, we send a reply to the recipient with their key.

QuorumProposalResponseRecv(Proposal<TYPES, QuorumProposalWrapper<TYPES>>)

A quorum proposal was requested by a node for a view.

DaProposalSend(Proposal<TYPES, DaProposal2<TYPES>>, TYPES::SignatureKey)

Send a DA proposal to the DA committee; emitted by the DA leader (which is the same node as the leader of view v + 1) in the DA task

DaVoteSend(DaVote2<TYPES>)

Send a DA vote to the DA leader; emitted by DA committee members in the DA task after seeing a valid DA proposal

QcFormed(Either<QuorumCertificate<TYPES>, TimeoutCertificate<TYPES>>)

The next leader has collected enough votes to form a QC; emitted by the next leader in the consensus task; an internal event only

Qc2Formed(Either<QuorumCertificate2<TYPES>, TimeoutCertificate2<TYPES>>)

The next leader has collected enough votes to form a QC; emitted by the next leader in the consensus task; an internal event only

NextEpochQc2Formed(Either<NextEpochQuorumCertificate2<TYPES>, TimeoutCertificate<TYPES>>)

The next leader has collected enough votes from the next epoch nodes to form a QC; emitted by the next leader in the consensus task; an internal event only

DacSend(DaCertificate2<TYPES>, TYPES::SignatureKey)

The DA leader has collected enough votes to form a DAC; emitted by the DA leader in the DA task; sent to the entire network via the networking task

ViewChange(TYPES::View, Option<TYPES::Epoch>)

The current view has changed; emitted by the replica in the consensus task or replica in the view sync task; received by almost all other tasks

ViewSyncTimeout(TYPES::View, u64, ViewSyncPhase)

Timeout for the view sync protocol; emitted by a replica in the view sync task

ViewSyncPreCommitVoteRecv(ViewSyncPreCommitVote2<TYPES>)

Receive a ViewSyncPreCommitVote from the network; received by a relay in the view sync task

ViewSyncCommitVoteRecv(ViewSyncCommitVote2<TYPES>)

Receive a ViewSyncCommitVote from the network; received by a relay in the view sync task

ViewSyncFinalizeVoteRecv(ViewSyncFinalizeVote2<TYPES>)

Receive a ViewSyncFinalizeVote from the network; received by a relay in the view sync task

ViewSyncPreCommitVoteSend(ViewSyncPreCommitVote2<TYPES>)

Send a ViewSyncPreCommitVote from the network; emitted by a replica in the view sync task

ViewSyncCommitVoteSend(ViewSyncCommitVote2<TYPES>)

Send a ViewSyncCommitVote from the network; emitted by a replica in the view sync task

ViewSyncFinalizeVoteSend(ViewSyncFinalizeVote2<TYPES>)

Send a ViewSyncFinalizeVote from the network; emitted by a replica in the view sync task

ViewSyncPreCommitCertificateRecv(ViewSyncPreCommitCertificate2<TYPES>)

Receive a ViewSyncPreCommitCertificate from the network; received by a replica in the view sync task

ViewSyncCommitCertificateRecv(ViewSyncCommitCertificate2<TYPES>)

Receive a ViewSyncCommitCertificate from the network; received by a replica in the view sync task

ViewSyncFinalizeCertificateRecv(ViewSyncFinalizeCertificate2<TYPES>)

Receive a ViewSyncFinalizeCertificate from the network; received by a replica in the view sync task

ViewSyncPreCommitCertificateSend(ViewSyncPreCommitCertificate2<TYPES>, TYPES::SignatureKey)

Send a ViewSyncPreCommitCertificate from the network; emitted by a relay in the view sync task

ViewSyncCommitCertificateSend(ViewSyncCommitCertificate2<TYPES>, TYPES::SignatureKey)

Send a ViewSyncCommitCertificate from the network; emitted by a relay in the view sync task

ViewSyncFinalizeCertificateSend(ViewSyncFinalizeCertificate2<TYPES>, TYPES::SignatureKey)

Send a ViewSyncFinalizeCertificate from the network; emitted by a relay in the view sync task

ViewSyncTrigger(TYPES::View)

Trigger the start of the view sync protocol; emitted by view sync task; internal trigger only

Timeout(TYPES::View, Option<TYPES::Epoch>)

A consensus view has timed out; emitted by a replica in the consensus task; received by the view sync task; internal event only

TransactionsRecv(Vec<TYPES::Transaction>)

Receive transactions from the network

TransactionSend(TYPES::Transaction, TYPES::SignatureKey)

Send transactions to the network

SendPayloadCommitmentAndMetadata(VidCommitment, BuilderCommitment, <TYPES::BlockPayload as BlockPayload<TYPES>>::Metadata, TYPES::View, Vec1<BuilderFee<TYPES>>, Option<TYPES::AuctionResult>)

Event to send block payload commitment and metadata from DA leader to the quorum; internal event only

BlockRecv(PackedBundle<TYPES>)

Event when the transactions task has sequenced transactions. Contains the encoded transactions, the metadata, and the view number

VidDisperseSend(Proposal<TYPES, VidDisperse<TYPES>>, TYPES::SignatureKey)

Send VID shares to VID storage nodes; emitted by the DA leader

Like HotShotEvent::DaProposalSend.

VidShareRecv(TYPES::SignatureKey, Proposal<TYPES, VidDisperseShare2<TYPES>>)

Vid disperse share has been received from the network; handled by the consensus task

Like HotShotEvent::DaProposalRecv.

VidShareValidated(Proposal<TYPES, VidDisperseShare2<TYPES>>)

VID share data is validated.

UpgradeProposalRecv(Proposal<TYPES, UpgradeProposal<TYPES>>, TYPES::SignatureKey)

Upgrade proposal has been received from the network

UpgradeProposalSend(Proposal<TYPES, UpgradeProposal<TYPES>>, TYPES::SignatureKey)

Upgrade proposal has been sent to the network

UpgradeVoteRecv(UpgradeVote<TYPES>)

Upgrade vote has been received from the network

UpgradeVoteSend(UpgradeVote<TYPES>)

Upgrade vote has been sent to the network

UpgradeCertificateFormed(UpgradeCertificate<TYPES>)

Upgrade certificate has been sent to the network

QuorumProposalPreliminarilyValidated(Proposal<TYPES, QuorumProposalWrapper<TYPES>>)

A quorum proposal has been preliminarily validated. The preliminary checks include:

The proposal is not for an old view

The proposal has been correctly signed by the leader of the current view

The justify QC is valid

VidRequestSend(DataRequest<TYPES>, TYPES::SignatureKey, TYPES::SignatureKey)

Send a VID request to the network; emitted to on of the members of DA committee. Includes the data request, node’s public key and signature as well as public key of DA committee who we want to send to.

VidRequestRecv(DataRequest<TYPES>, TYPES::SignatureKey)

Receive a VID request from the network; Received by a node in the DA committee. Includes the data request and nodes public key.

VidResponseSend(TYPES::SignatureKey, TYPES::SignatureKey, Proposal<TYPES, VidDisperseShare2<TYPES>>)

Send a VID response to the network; emitted to the sending node. Includes nodes public key, recipient public key, and vid disperse

Tuple Fields

0: TYPES::SignatureKey

Sender key

1: TYPES::SignatureKey

Recipient key

2: Proposal<TYPES, VidDisperseShare2<TYPES>>

VidResponseRecv(TYPES::SignatureKey, Proposal<TYPES, VidDisperseShare2<TYPES>>)

Receive a VID response from the network; received by the node that triggered the VID request.

HighQcRecv(QuorumCertificate2<TYPES>, TYPES::SignatureKey)

A replica send us a High QC

HighQcSend(QuorumCertificate2<TYPES>, TYPES::SignatureKey, TYPES::SignatureKey)

Send our HighQc to the next leader, should go to the same leader as our vote

Implementations

impl<TYPES: NodeType> HotShotEvent<TYPES>

pub fn view_number(&self) -> Option<TYPES::View>

Return the view number for a hotshot event if present