Module hotshot_task_impls::helpers

Helper functions used by any task

Structs

LeafChainTraversalOutcome

Helper type to give names and to the output values of the leaf chain traversal operation.

Constants

DONT_SEND_VIEW_CHANGE_EVENT

Constant which tells update_view to not send a view change event when called.

SEND_VIEW_CHANGE_EVENT 🔒

Constant which tells update_view to send a view change event when called.

Traits

AnyhowTracing

Utilities to print anyhow logs.

Functions

broadcast_event

Helper function to send events and log errors

cancel_task

Cancel a task

decide_from_proposal

Ascends the leaf chain by traversing through the parent commitments of the proposal. We begin by obtaining the parent view, and if we are in a chain (i.e. the next view from the parent is one view newer), then we begin attempting to form the chain. This is a direct impl from HotStuff section 5:

fetch_proposal 🔒

Trigger a request to the network for a proposal for a view and wait for the response or timeout.

parent_leaf_and_state 🔒

Gets the parent leaf and state from the parent of a proposal, returning an anyhow::Error if not.

update_view 🔒

Update the view if it actually changed, takes a mutable reference to the cur_view and the timeout_task which are updated during the operation of the function.

validate_proposal_safety_and_liveness

Validate the state and safety and liveness of a proposal then emit a QuorumProposalValidated event.

validate_proposal_view_and_certs

Validates, from a given proposal that the view that it is being submitted for is valid when compared to cur_view which is the highest proposed view (so far) for the caller. If the proposal is for a view that’s later than expected, that the proposal includes a timeout or view sync certificate.