hotshot_task_impls::helpers

Function wait_for_second_vid_share

pub async fn wait_for_second_vid_share<TYPES: NodeType>(
    target_epoch: Option<TYPES::Epoch>,
    vid_share: &Proposal<TYPES, VidDisperseShare<TYPES>>,
    da_cert: &DaCertificate2<TYPES>,
    consensus: &OuterConsensus<TYPES>,
    timeout: u64,
    view_start_time: Instant,
    receiver: &Receiver<Arc<HotShotEvent<TYPES>>>,
) -> Result<Proposal<TYPES, VidDisperseShare<TYPES>>>

Gets the second VID share, the current or the next epoch accordingly, from the shared consensus state; makes sure it corresponds to the given DA certificate; if it’s not yet available, waits for it with the given timeout.