Type Alias hotshot_task_impls::network::test::ModifierClosure

pub type ModifierClosure<TYPES> = dyn Fn(&mut <TYPES as NodeType>::SignatureKey, &mut MessageKind<TYPES>, &mut TransmitType<TYPES>, &<TYPES as NodeType>::Membership) + Send + Sync;

A dynamic type alias for a function that takes the result of NetworkEventTaskState::parse_event and changes it before transmitting on the network.