hotshot_task_impls::quorum_proposal_recv
Module handlers

Event handlers for this task.

Functions

handle_quorum_proposal_recv 🔒

Handles the QuorumProposalRecv event by first validating the cert itself for the view, and then updating the states, which runs when the proposal cannot be found in the internal state map.

spawn_fetch_proposal 🔒

Spawn a task which will fire a request to get a proposal, and store it.

validate_proposal_liveness 🔒

Update states in the event that the parent state is not found for a given proposal.