hotshot_task_impls::quorum_vote

Module handlers

Event handlers for QuorumProposalValidated.

Functions

get_computed_drb_result 🔒

Store the DRB result from the computation task to the shared results table.

handle_quorum_proposal_validated 🔒

Handles the QuorumProposalValidated event.

store_drb_result 🔒

Store the DRB result for the next epoch if we received it in a decided leaf.

submit_vote 🔒

Submits the QuorumVoteSend event if all the dependencies are met.

update_shared_state 🔒

Updates the shared consensus state with the new voting data.

verify_drb_result 🔒

Verify the DRB result from the proposal for the next epoch if this is the last block of the current epoch.