hotshot_task_impls

Module response

Task which responses to requests from the network

Structs

NetworkResponseState

Task state for the Network Request Task. The task is responsible for handling requests sent to this node by the network. It will validate the sender, parse the request, and try to find the data request in the consensus stores.

Constants

TXNS_TIMEOUT 🔒

Time to wait for txns before sending ResponseMessage::NotFound

Functions

run_response_task

Spawn the network response task to handle incoming request for data from other nodes. It will shutdown when it gets HotshotEvent::Shutdown on the event_stream arg.

valid_signature 🔒

Check the signature