Struct hotshot_task_impls::transactions::TransactionTaskState

pub struct TransactionTaskState<TYPES: NodeType, I: NodeImplementation<TYPES>, V: Versions> {

    pub builder_timeout: Duration,
    pub output_event_stream: Sender<Event<TYPES>>,
    pub cur_view: TYPES::View,
    pub cur_epoch: Option<TYPES::Epoch>,
    pub consensus: OuterConsensus<TYPES>,
    pub membership: Arc<RwLock<TYPES::Membership>>,
    pub builder_clients: Vec<BuilderClient<TYPES>>,
    pub public_key: TYPES::SignatureKey,
    pub private_key: <TYPES::SignatureKey as SignatureKey>::PrivateKey,
    pub instance_state: Arc<TYPES::InstanceState>,
    pub id: u64,
    pub upgrade_lock: UpgradeLock<TYPES, V>,
    pub auction_results_provider: Arc<I::AuctionResultsProvider>,
    pub fallback_builder_url: Url,
    pub epoch_height: u64,

}

Tracks state of a Transaction task

Fields
builder_timeout: Duration

The state’s api

output_event_stream: Sender<Event<TYPES>>

Output events to application

cur_view: TYPES::View

View number this view is executing in.

cur_epoch: Option<TYPES::Epoch>

Epoch number this node is executing in.

consensus: OuterConsensus<TYPES>

Reference to consensus. Leader will require a read lock on this.

membership: Arc<RwLock<TYPES::Membership>>

Membership for the quorum

builder_clients: Vec<BuilderClient<TYPES>>

Builder 0.1 API clients

public_key: TYPES::SignatureKey

This Nodes Public Key

private_key: <TYPES::SignatureKey as SignatureKey>::PrivateKey

Our Private Key

instance_state: Arc<TYPES::InstanceState>

InstanceState

id: u64

This state’s ID

upgrade_lock: UpgradeLock<TYPES, V>

Lock for a decided upgrade

auction_results_provider: Arc<I::AuctionResultsProvider>

auction results provider

fallback_builder_url: Url

fallback builder url

epoch_height: u64

Number of blocks in an epoch, zero means there are no epochs

Implementations

impl<TYPES: NodeType, I: NodeImplementation<TYPES>, V: Versions> TransactionTaskState<TYPES, I, V>

pub async fn handle_view_change( &mut self, event_stream: &Sender<Arc<HotShotEvent<TYPES>>>, block_view: TYPES::View, block_epoch: Option<TYPES::Epoch>, ) -> Option<HotShotTaskCompleted>

handle view change decide legacy or not

pub async fn handle_view_change_legacy( &mut self, event_stream: &Sender<Arc<HotShotEvent<TYPES>>>, block_view: TYPES::View, block_epoch: Option<TYPES::Epoch>, ) -> Option<HotShotTaskCompleted>

legacy view change handler

async fn produce_block_marketplace( &mut self, block_view: TYPES::View, block_epoch: Option<TYPES::Epoch>, task_start_time: Instant, ) -> Result<PackedBundle<TYPES>>

Produce a block by fetching auction results from the solver and bundles from builders.

Errors

Returns an error if the solver cannot be contacted, or if none of the builders respond.

pub async fn null_block( &self, block_view: TYPES::View, block_epoch: Option<TYPES::Epoch>, version: Version, ) -> Option<PackedBundle<TYPES>>

Produce a null block

pub async fn handle_view_change_marketplace( &mut self, event_stream: &Sender<Arc<HotShotEvent<TYPES>>>, block_view: TYPES::View, block_epoch: Option<TYPES::Epoch>, ) -> Option<HotShotTaskCompleted>

marketplace view change handler

pub async fn handle_view_change_epochs( &mut self, event_stream: &Sender<Arc<HotShotEvent<TYPES>>>, block_view: TYPES::View, block_epoch: Option<TYPES::Epoch>, ) -> Option<HotShotTaskCompleted>

epochs view change handler

pub async fn handle( &mut self, event: Arc<HotShotEvent<TYPES>>, event_stream: Sender<Arc<HotShotEvent<TYPES>>>, ) -> Result<()>

main task event handler

async fn last_vid_commitment_retry( &self, block_view: TYPES::View, task_start_time: Instant, ) -> Result<(TYPES::View, VidCommitment)>

Get VID commitment for the last successful view before block_view. Returns None if we don’t have said commitment recorded.

async fn last_vid_commitment( &self, block_view: TYPES::View, ) -> Result<(TYPES::View, VidCommitment)>

Get VID commitment for the last successful view before block_view. Returns None if we don’t have said commitment recorded.

async fn wait_for_block( &self, block_view: TYPES::View, ) -> Option<BuilderResponse<TYPES>>

async fn get_available_blocks( &self, parent_comm: VidCommitment, view_number: TYPES::View, parent_comm_sig: &<<TYPES as NodeType>::SignatureKey as SignatureKey>::PureAssembledSignatureType, ) -> Vec<(AvailableBlockInfo<TYPES>, usize)>

Query the builders for available blocks. Queries only fraction of the builders based on the response time.

async fn block_from_builder( &self, parent_comm: VidCommitment, view_number: TYPES::View, parent_comm_sig: &<<TYPES as NodeType>::SignatureKey as SignatureKey>::PureAssembledSignatureType, ) -> Result<BuilderResponse<TYPES>>

Get a block from builder. Queries the sufficiently fast builders for available blocks and chooses the one with the best fee/byte ratio, re-trying with the next best one in case of failure.

Errors

If none of the builder reports any available blocks or claiming block fails for all of the builders.