hotshot_task_impls
Module vote_collection

Generic task for collecting votes

Structs

AccumulatorInfo

Info needed to create a vote accumulator task

VoteCollectionTaskState

Task state for collecting votes of one type and emitting a certificate

Traits

AggregatableVote

Describes the functions a vote must implement for it to be aggregatable by the generic vote collection task

HandleVoteEvent

Trait for types which will handle a vote event.

Functions

create_vote_accumulator

Generic function for spawning a vote task. Returns the event stream id of the spawned task if created

handle_vote

A helper function that handles a vote regardless whether it’s the first vote in the view or not.

Type Aliases

DaVoteState 🔒

Alias for DA vote accumulator

NextEpochQuorumVoteState 🔒

Alias for Quorum vote accumulator

QuorumVoteState 🔒

Alias for Quorum vote accumulator

TimeoutVoteState 🔒

Alias for Timeout vote accumulator

UpgradeVoteState 🔒

Alias for upgrade vote accumulator

ViewSyncCommitVoteState 🔒

Alias for View Sync Commit vote accumulator

ViewSyncFinalizeVoteState 🔒

Alias for View Sync Finalize vote accumulator

ViewSyncPreCommitState 🔒

Alias for View Sync Pre Commit vote accumulator

VoteCollectorsMap

Alias for a map of Vote Collectors