Crate hotshot_task_impls

The consensus layer for hotshot. This currently implements sequencing consensus in an event driven way

Modules

builder

Implementations for builder client Should contain builder task in the future

consensus

the task which implements the main parts of consensus

da

The task which implements the main parts of data availability.

events

Defines the events passed between tasks

harness

Defines the types to run unit tests for a task.

helpers

Helper functions used by any task

network

The task which implements the network.

response

Task which responsds to requests from the network

transactions

The task which implements all transaction handling

upgrade

Task for handling upgrades

vid

The task which implements verifiable information dispersal

view_sync

The task which implements view synchronization

vote

Generic task for collecting votes