Crate hotshot_task_impls

The consensus layer for hotshot. This currently implements sequencing consensus in an event driven way

Modules

builder

Implementations for builder client Should contain builder task in the future

consensus

The task which implements the core state logic of consensus.

da

The task which implements the main parts of data availability.

events

Defines the events passed between tasks

harness

Defines the types to run unit tests for a task.

helpers

Helper functions used by any task

network

The task which implements the network.

quorum_proposal

Task for handling logic for quorum proposals

quorum_proposal_recv

Task for handling QuorumProposalRecv events

quorum_vote

The task which handles the logic for the quorum vote.

request

Task for requesting the network for things

response

Task which responses to requests from the network

rewind

Task for storing and replaying all received tasks by a node

transactions

The task which implements all transaction handling

upgrade

Task for handling upgrades

vid

The task which implements verifiable information dispersal

view_sync

The task which implements view synchronization

vote_collection

Generic task for collecting votes