hotshot_task_impls::quorum_vote
Module handlers

Event handlers for QuorumProposalValidated.

Functions

handle_quorum_proposal_validated 🔒

Handles the QuorumProposalValidated event.

start_drb_task 🔒

Start the DRB computation task for the next epoch.

store_and_get_computed_drb_result 🔒

Store the DRB result from the computation task to the shared results table.

store_drb_seed_and_result 🔒

Store the DRB seed two epochs in advance and the computed or received DRB result for next epoch.

submit_vote 🔒

Submits the QuorumVoteSend event if all the dependencies are met.

update_shared_state 🔒

Updates the shared consensus state with the new voting data.

verify_drb_result 🔒

Verify the DRB result from the proposal for the next epoch if this is the last block of the current epoch.