hotshot_task_impls
Module transactions

The task which implements all transaction handling

Structs

BuilderResponse

Builder Provided Responses

TransactionTaskState

Tracks state of a Transaction task

Constants

BUILDER_ADDITIONAL_TIME_MULTIPLIER 🔒

Multiplier for extra time to give to the second batch of builders

BUILDER_MAIN_BATCH_CUTOFF 🔒

Time the first batch of builders has to respond

BUILDER_MAIN_BATCH_THRESHOLD_DIVIDEND 🔒

Proportion of builders queried in first batch, dividend

BUILDER_MAIN_BATCH_THRESHOLD_DIVISOR 🔒

Proportion of builders queried in the first batch, divisor

BUILDER_MINIMUM_QUERY_TIME 🔒

Minimum amount of time allotted to both batches, cannot be cut shorter if the first batch responds extremely fast.

RETRY_DELAY 🔒

Delay between re-tries on unsuccessful calls