Trait hotshot_task_impls::vote::AggregatableVote

pub trait AggregatableVote<TYPES: NodeType, VOTE: Vote<TYPES>, CERT: Certificate<TYPES, Voteable = VOTE::Commitment>> {
    // Required methods
    fn get_leader(&self, membership: &TYPES::Membership) -> TYPES::SignatureKey;
    fn make_cert_event(
        certificate: CERT,
        key: &TYPES::SignatureKey
    ) -> HotShotEvent<TYPES>;
}

Describes the functions a vote must implement for it to be aggregatable by the generic vote collection task

Required Methods

fn get_leader(&self, membership: &TYPES::Membership) -> TYPES::SignatureKey

return the leader for this votes

fn make_cert_event( certificate: CERT, key: &TYPES::SignatureKey ) -> HotShotEvent<TYPES>

return the Hotshot event for the completion of this CERT