hotshot_task_impls::vote_collection
Trait AggregatableVote

pub trait AggregatableVote<TYPES: NodeType, VOTE: Vote<TYPES>, CERT: Certificate<TYPES, VOTE::Commitment, Voteable = VOTE::Commitment>> {
    // Required methods
    fn leader(
        &self,
        membership: &TYPES::Membership,
        epoch: Option<TYPES::Epoch>,
    ) -> Result<TYPES::SignatureKey>;
    fn make_cert_event(
        certificate: CERT,
        key: &TYPES::SignatureKey,
    ) -> HotShotEvent<TYPES>;
}

Describes the functions a vote must implement for it to be aggregatable by the generic vote collection task

Required Methods

fn leader( &self, membership: &TYPES::Membership, epoch: Option<TYPES::Epoch>, ) -> Result<TYPES::SignatureKey>

return the leader for this votes

Errors

if the leader cannot be calculated

fn make_cert_event( certificate: CERT, key: &TYPES::SignatureKey, ) -> HotShotEvent<TYPES>

return the Hotshot event for the completion of this CERT

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.