Trait hotshot_task_impls::vote_collection::HandleVoteEvent

pub trait HandleVoteEvent<TYPES, VOTE, CERT>
where
    TYPES: NodeType,
    VOTE: Vote<TYPES> + AggregatableVote<TYPES, VOTE, CERT>,
    CERT: Certificate<TYPES, VOTE::Commitment, Voteable = VOTE::Commitment> + Debug,
{
    // Required methods
    fn handle_vote_event<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        event: Arc<HotShotEvent<TYPES>>,
        sender: &'life1 Sender<Arc<HotShotEvent<TYPES>>>,
    ) -> Pin<Box<dyn Future<Output = Result<Option<CERT>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn filter(event: Arc<HotShotEvent<TYPES>>) -> bool;
}

Trait for types which will handle a vote event.

Required Methods

fn handle_vote_event<'life0, 'life1, 'async_trait>( &'life0 mut self, event: Arc<HotShotEvent<TYPES>>, sender: &'life1 Sender<Arc<HotShotEvent<TYPES>>>, ) -> Pin<Box<dyn Future<Output = Result<Option<CERT>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Handle a vote event

Errors

Returns an error if we fail to handle the vote

fn filter(event: Arc<HotShotEvent<TYPES>>) -> bool

Event filter to use for this event

Object Safety

This trait is not object safe.